Move signal module out of global scope
Because we're (ab)using `PYTHONPATH` to bring UltiMaker modules into
scope, the `signal` module of `libSmeagol` is masking the stdlib
`signal` module. This is a problem when third-party libraries, such as
*pytest*, try to `import sigal`.

As a workaround, move the `libSmeagol.signal` module into a `util`
subpackage.

(It looks like `libSmeagol.signal` is only used within `libSmeagol`
itself, so this move should not affect any users of `libSmeagol`. Users
of `libSmeagol` typically vendor their own versions of `signal`, for
better or worse.)
