+++ libSmeagol/util/signal.py
